// src/modules/issue/issue.interface.ts
export interface TIssue {
  id?: number;
  title: string;
  description: string;
  type: 'bug' | 'feature_request';
  status?: 'open' | 'in_progress' | 'resolved';
  reporter_id: number;
  created_at?: Date;
  updated_at?: Date;
}

export interface TIssueQueryParams {
  sort?: 'newest' | 'oldest';
  type?: 'bug' | 'feature_request';
  status?: 'open' | 'in_progress' | 'resolved';
}